[ls-static-site] feat: move func to own file
